requirements.txt
transformers
faiss-cpu
streamlit
fastapi
uvicorn[standard]
requests
